move cli to folder structure
