+++ template/src/{{ project_name }}/__main__.py
@@ -1,4 +1,4 @@
-from ._cli import app
+from .cli import app
 
 
 app()
